Added actual module code, partially working
This is a Textadept module for creating generic single-line views
that can be programmatically controlled to display arbitrary
information. The initial purpose was to be a statusbar alternative,
but it has been built generically so it could also be attached to
the top of a buffer to be used as a status area, or click events
could be attached to it to be used as a tab bar, file menu, etc.

Currently the best way to demo it is to open a horizontal split at the top level, focus the lower view, open a command entry, and type
textbar = require("textbar")
textbar(view)
textbar.replace_statusbar()
To shrink the view and hijack statusbar functionality. The current
major bug is that the RESIZE event hook is closing the command entry
preventing text from being entered. I think I can circumvent it by
adding a function that gets called to the key combo/menu that opens
the command entry and turns off the resize event handler.