Fix command entry handling, clean up interface
I was able to fix the issue where command entry handling broke
due to resize and view switch events. For now we just disable
update() while the command_entry is open. This could be addressed
in the future if I get around to writing that command entry
replacement with no special handling. I tried to fix the find
dialog losing focus as well but it looks like ui.find.focus()
won't reselect the find field if called from an event handler for
some reason. This is in a usable state now, but it's still a little
fragile and some edge cases will break it.