Fix Find/Replace bug, removed view switching code
So the description in the Textadept API doc of the relationship
between buffer and view is straight up misleading. I traced it today
and discovered that you can totally update just about everything in
a non-active view as long as you are referencing the correct table.
90% of the time it should be view, not buffer. After I realized that
I switched which table I was referencing and was able to remove all
of the view switching code, which meant I was also able to remove all
the special handling of the command entry and fix the Find dialog.